requirements.txt
Flask==3.1.3
PyGitHub==2.9.0
github-webhook==1.0.4